Adressed comments
Moved macOS native pointer support to separate file, passing proper event source to PointerRoutedEventArgs, supporting keyboard modifiers
