+++ src/Uno.UI/UI/Xaml/UIElement.macOS.cs
@@ -41,22 +41,7 @@ namespace Windows.UI.Xaml
 		public UIElement()
 		{
 			InitializePointers();
-		}
-
-		public override void UpdateTrackingAreas()
-		{
-			if (_trackingArea != null)
-			{
-				RemoveTrackingArea(_trackingArea);
-			}
-			var options =
-				NSTrackingAreaOptions.MouseEnteredAndExited |
-				NSTrackingAreaOptions.MouseMoved |
-				NSTrackingAreaOptions.ActiveInKeyWindow;
-			
-			_trackingArea = new NSTrackingArea(this.Bounds, options, this, null);
-			AddTrackingArea(_trackingArea);
-		}
+		}		
 		
 		internal bool ClippingIsSetByCornerRadius { get; set; } = false;
 
@@ -252,235 +237,6 @@ namespace Windows.UI.Xaml
 		}
 #endregion
 
-		internal void RaiseTapped(TappedRoutedEventArgs args) => LogUnRegisterPointerReleasedNotImplemented();
-
-		public override void MouseDown(NSEvent evt)
-		{
-			try
-			{
-				var pointerEventIsHandledInManaged = false;
-
-				if (evt.IsTouchInView(this))
-				{
-					IsPointerPressed = true;
-					IsPointerOver = true;
-
-					// evt.AllTouches raises a invalid selector exception
-					var args = new PointerRoutedEventArgs(null, evt)
-					{
-						CanBubbleNatively = true,
-						OriginalSource = this
-					};
-
-					pointerEventIsHandledInManaged = RaisePointerEvent(PointerPressedEvent, args);
-				}
-
-				if (!pointerEventIsHandledInManaged)
-				{
-					// Bubble up the event natively
-					base.MouseDown(evt);
-				}
-			}
-			catch (Exception e)
-			{
-				Application.Current.RaiseRecoverableUnhandledException(e);
-			}
-		}
-
-		public override void MouseMoved(NSEvent evt)
-		{
-			try
-			{
-				var pointerEventIsHandledInManaged = false;
-
-				if (evt.IsTouchInView(this))
-				{
-					IsPointerPressed = true;
-					IsPointerOver = true;
-
-					// evt.AllTouches raises a invalid selector exception
-					var args = new PointerRoutedEventArgs(null, evt)
-					{
-						CanBubbleNatively = true,
-						OriginalSource = this
-					};
-					
-					pointerEventIsHandledInManaged = RaisePointerEvent(PointerMovedEvent, args);
-				}
-
-				if (!pointerEventIsHandledInManaged)
-				{
-					// Bubble up the event natively
-					base.MouseMoved(evt);
-				}
-			}
-			catch (Exception e)
-			{
-				Application.Current.RaiseRecoverableUnhandledException(e);
-			}
-		}
-
-		public override void MouseEntered(NSEvent evt)
-		{
-			try
-			{
-				var pointerEventIsHandledInManaged = false;
-
-				if (evt.IsTouchInView(this))
-				{
-					IsPointerPressed = true;
-					IsPointerOver = true;
-
-					// evt.AllTouches raises a invalid selector exception
-					var args = new PointerRoutedEventArgs(null, evt)
-					{
-						CanBubbleNatively = true,
-						OriginalSource = this
-					};
-
-					pointerEventIsHandledInManaged = RaisePointerEvent(PointerEnteredEvent, args);
-				}
-
-				if (!pointerEventIsHandledInManaged)
-				{
-					// Bubble up the event natively
-					base.MouseEntered(evt);
-				}
-			}
-			catch (Exception e)
-			{
-				Application.Current.RaiseRecoverableUnhandledException(e);
-			}
-		}
-
-		public override void MouseExited(NSEvent evt)
-		{
-			try
-			{
-				var pointerEventIsHandledInManaged = false;
-
-				if (evt.IsTouchInView(this))
-				{
-					IsPointerPressed = true;
-					IsPointerOver = true;
-
-					// evt.AllTouches raises a invalid selector exception
-					var args = new PointerRoutedEventArgs(null, evt)
-					{
-						CanBubbleNatively = true,
-						OriginalSource = this
-					};
-
-					pointerEventIsHandledInManaged = RaisePointerEvent(PointerExitedEvent, args);
-				}
-
-				if (!pointerEventIsHandledInManaged)
-				{
-					// Bubble up the event natively
-					base.MouseExited(evt);
-				}
-			}
-			catch (Exception e)
-			{
-				Application.Current.RaiseRecoverableUnhandledException(e);
-			}
-		}
-
-		public override void MouseUp(NSEvent evt)
-		{
-			try
-			{
-				var wasPointerOver = IsPointerOver;
-				IsPointerOver = evt.IsTouchInView(this);
-
-				// Call entered/exited one last time
-				// evt.AllTouches raises a invalid selector exception
-				var args = new PointerRoutedEventArgs(null, evt)
-				{
-					CanBubbleNatively = true,
-					OriginalSource = this
-				};
-
-				var pointerEventIsHandledInManaged = false;
-
-				if (!wasPointerOver && IsPointerOver)
-				{
-					pointerEventIsHandledInManaged = RaisePointerEvent(PointerEnteredEvent, args);
-				}
-				else if (wasPointerOver && !IsPointerOver)
-				{
-					pointerEventIsHandledInManaged = RaisePointerEvent(PointerExitedEvent, args);
-				}
-
-				if (IsPointerCaptured || IsPointerOver)
-				{
-					args.Handled = false; // reset as unhandled
-					pointerEventIsHandledInManaged = RaisePointerEvent(PointerReleasedEvent, args) || pointerEventIsHandledInManaged;
-				}
-
-				if (IsPointerCaptured)
-				{
-					args.Handled = false; // reset as unhandled
-					pointerEventIsHandledInManaged = RaisePointerEvent(PointerCaptureLostEvent, args) || pointerEventIsHandledInManaged;
-				}
-
-				if (!pointerEventIsHandledInManaged)
-				{
-					// Bubble up the event natively
-					base.MouseUp(evt);
-				}
-
-				IsPointerPressed = false;
-				IsPointerOver = false;
-			}
-			catch (Exception e)
-			{
-				Application.Current.RaiseRecoverableUnhandledException(e);
-			}
-		}
-
-		public override void MouseDragged(NSEvent evt)
-		{
-			try
-			{
-				var wasPointerOver = IsPointerOver;
-				IsPointerOver = evt.IsTouchInView(this);
-
-				var pointerEventIsHandledInManaged = false;
-
-				// evt.AllTouches raises a invalid selector exception
-				var args = new PointerRoutedEventArgs(null, evt)
-				{
-					CanBubbleNatively = true,
-					OriginalSource = this
-				};
-
-				if (IsPointerCaptured || IsPointerOver)
-				{
-					pointerEventIsHandledInManaged = RaisePointerEvent(PointerMovedEvent, args);
-				}
-
-				if (!wasPointerOver && IsPointerOver)
-				{
-					args.Handled = false; // reset as unhandled
-					pointerEventIsHandledInManaged = RaisePointerEvent(PointerEnteredEvent, args) || pointerEventIsHandledInManaged;
-				}
-				else if (wasPointerOver && !IsPointerOver)
-				{
-					args.Handled = false; // reset as unhandled
-					pointerEventIsHandledInManaged = RaisePointerEvent(PointerExitedEvent, args) || pointerEventIsHandledInManaged;
-				}
-
-				if (!pointerEventIsHandledInManaged)
-				{
-					// Bubble up the event natively
-					base.MouseDragged(evt);
-				}
-			}
-			catch (Exception e)
-			{
-				Application.Current.RaiseRecoverableUnhandledException(e);
-			}
-		}
+		internal void RaiseTapped(TappedRoutedEventArgs args) => LogUnRegisterPointerReleasedNotImplemented();		
 	}
 }
